Todas las paginas responsive

+++ src/main/java/np/com/proyecto/dao/RolDao.java
@@ -3,6 +3,6 @@ package np.com.proyecto.dao;
 import np.com.proyecto.domain.Rol;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolDao extends JpaRepository<Rol, Long> {
+public interface RolDao extends JpaRepository<Rol, Integer> {
     
 }
